token validation in user and contact

+++ 08contactmanagement/controller/contact.js
@@ -4,7 +4,7 @@ const Contact = require("../models/contact");
 //@route GET /api/contacts
 //@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
@@ -21,7 +21,8 @@ const createContact = asyncHandler(async (req, res) => {
   const contact = await Contact.create({
     name,
     email,
-    phone
+    phone,
+    user_id: req.user.id
   });
 
   res.status(201).json(contact);
@@ -52,6 +53,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if(contact.user_id.toString() != req.user.id) {
+    req.status = 403
+    throw new Error('User is not authorized to update')
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -70,6 +76,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+
+  if(contact.user_id.toString() != req.user.id) {
+    req.status = 403
+    throw new Error('User is not authorized to delte')
+  }
+
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
